refactor(event-landing): drop EventLandingPage wrapper and add address data
- Wrap component with Layout including page title/description/image metadata
- Include address in structured data JSON-LD when available
- Remove the now-unnecessary EventLandingPage component
- Simplify landing page files to import EventLanding directly
